+++ infrastructure/commons/external_dns/namespace.tf
@@ -1,0 +1,8 @@
+resource "kubernetes_namespace" "external_dns" {
+  metadata {
+    name = var.external_dns_namespace
+    labels = {
+      name = var.external_dns_namespace
+    }
+  }
+}