+++ nullplatform/account/variables.tf
@@ -1,5 +1,5 @@
 variable "nullplatform_accounts" {
-  description = "Map of Nullplatform accounts to create with their configuration settings"
+  description = "A map of nullplatform accounts to create with their configuration settings"
   type = map(object({
     name                = string
     repository_prefix   = optional(string, "poc-account")
@@ -10,6 +10,6 @@ variable "nullplatform_accounts" {
 
 variable "np_api_key" {
   type        = string
-  description = "The API token must be at the organization level"
+  description = "The nullplatform API key (must be at the organization level)"
 
 }