+++ nullplatform/agent/auth.tf
@@ -1,5 +1,5 @@
 ################################################################################
-# Nullplatform Agent API Key
+# Nullplatform agent API key
 ################################################################################
 
 # Create API key for agent authentication with required role grants
@@ -38,5 +38,10 @@ resource "nullplatform_api_key" "nullplatform_agent_api_key" {
   tags {
     key   = "managed-by"
     value = "IaC"
+  }
+
+  tags {
+    key   = "owner"
+    value = var.nrn
   }
 }