+++ nullplatform/base/providers.tf
@@ -15,7 +15,3 @@ terraform {
 
 
 
-provider "nullplatform" {
-
-  api_key = var.np_api_key
-}