+++ nullplatform/prometheus/variables.tf
@@ -5,31 +5,35 @@ variable "prometheus_namespace" {
 }
 
 variable "nrn" {
-  description = "Nullplatform Resource Name - unique identifier for resources"
+  description = "Nullplatform Resource Name — unique identifier for resources"
   type        = string
 }
 
 variable "np_api_key" {
-  description = "Nullplatform API key for authentication"
+  description = "nullplatform API key for authentication"
   type        = string
   sensitive   = true
 }
 
 variable "nullplatform_port" {
-  description = "Port number for Nullplatform service communication"
+  description = "Port number for nullplatform service communication"
   type        = number
   default     = 2021
 }
 
-variable "kubeconfig_path" {
-  description = "Path to the kubeconfig file for Kubernetes cluster access"
-  type        = string
-  default     = "~/.kube/config"
+variable "install_prometheus" {
+  description = "Activate the Prometheus installation"
+  type        = bool
+  default     = false
+
 }
 
-variable "kube_context" {
-  description = "Kubernetes context name to use from the kubeconfig file"
-  type        = string
-  default     = null
+variable "prometheus_url" {
+  type    = string
+  default = ""
 }
+variable "dimensions" {
+  default     = {}
+  description = "name of the dimensions"
 
+}